chore: Move BookSearch.tsx from /pages to /features/shelves/
- Moved the BookSearch.tsx file to /features/shelves/ to align with a modern React feature-based folder structure.
